bubble-sheet-generator/requirements.txt
flask==3.0.0
flask-cors==4.0.0
pika==1.3.2
python-dotenv==1.0.0
reportlab==4.0.7
qrcode==7.4.2
Pillow
requests==2.31.0
